[platform] add Percentile tests&tweak thread-safety
Add an argument to the Percentile constructor to optionally make
Percentile thread-safe by using a mutex to add points to a Percentile.

Instead of using a global lock on the PercentileRegistry for all
operations on the contained Percentiles. Use a global lock just for
registering Percentiles with the registry and thread-safe Percentile
instance. This avoids a global lock for recording every type of
percentile.