[olap] enable CPU only morsel from NVMe
Add adm-play to check that SSB queries can run. For now membrdcst is not
supported, so we can only use plans that do morsel joins and so do not
need to do a broadcast. Broadcasting data loaded from nvme is tricky
both logically and the implementation given the asynchronous IO.

Add some tests for router + mem-move when using NvmePlugin
Add a disabled test for membrdcast

Add adm-partition-ssb to split binary data files into partitions and
generate the metadata that the NvmePlugin expects.

Work around a weird issue between our allocators and structs containing
a std::function in io_uring

Add a nvme shaper for CPU only and morsel join plans