[Build] Fix release builds
We for some reason hard coded debug builds. This means NDEBUG was never
set.
Most changes are to fix the build when NDEBUG is set.
Importantly, `assert` compiles out with NDEGUG, things that should be
asserted in all builds should use `CHECK`. Things that should be
asserted in Debug, e.e. assertions that are slow or on the critical
path, builds should use `DCHECK`.