[platform] fix ProfileRegion
It seems that the string for __itt_event_create needed a longer
lifetime.

Additionally, there is some undocumented upperbound on the number of
events you can create with __itt_event_create, if you create thousands,
you will segfault in intels code. So we move to having a
ProfileRegionType which holds the event, and needs to be passed to
ProfileRegion. ProfileRegionType can generally be a static class member
where it is used if the name can be a static/constant string.

Also remove some unnecessary ifdefs cudaProfilerStart comes from
cuda_profiler_api.h not NVTX, and gpu-common already defines it for
NCUDA builds