[build] truly cursed cufile CMake
This is not the right way to do this, but this is a way to link to
cufile in cuda 12 while we use cuda 11.

The issue with the previous approach is when we try to find cufile from
cuda 12, it has the same include directory and link directory as
cuda 11, which overrides it. The way we use cuda right now sets global
include/link directories.