[olap] document affinitizers.hpp
Still a little unsure on getAvailableCUIndex, especially with respect
to CpuCoreAffinitizer. CpuCore returns a topology::core for getAvailableCU
and a CPUNumaNode index_in_topo for getAvailableCUIndex. I am not sure
why it does not return the index_in_topo for a cpu core. (topology does
not implement that currently, but it would be fairly trivial)