[adm] move gpu decompression to class
There are many small allocations that can be re-used, this is easiest
with a class interface instead of a function interface.

This commit implements a GpuDecompressor class for block of batch of
block decompression. To ease implementation, only a single async
decompression is supported. (Async with respect to the host)