[olap] fix bloomfilters
- get ptr types instead of array ptr types from IndexedSequence
expressions. This is a bit of a hack, but unsure of what the correct
solution is. We don't really define if IndexedSequence is a fixed size
or dynamic size container.
- Remove AVX512 intrinsics in bloomfilter_repack. (AMD Milan doesn't
have AVX512 so we would crash LLVM during compilation). I have not
verified performance / what code is generated on machines with AVX512
but on Milan this generates AVX2 vector loads
- Small amount of refactoring for readability
- Add bloom filter tests.