[adm/olap] for CPU only, do mem-move if data and compute affinity differ
Previously, if the block were on a different numa node from where the
query thread is executing it would never be moved. With this commit,
the data will be moved if do_transfer is true for the column the block
belongs to and the data is on a different CPU socket. This lets us
explore the trade-off of moving an entire block from one socket to
another, vs just accessing the values of that block as necessary during
pipeline execution across the interconnect.

Here we move if the data is on a different socket instead of on a
different numa node. Either option is reasonable, but I choose socket
level specifically because chiplet based CPUs like AMD EPYC have
multiple numa nodes per socket, and numa-to-numa BW on the same socket
is reasonably high compared to socket-to-socket BW.