[codegen] embed debug info in generated LLVM-IR
This commit adds the option to embed debug info into our generated
LLVM-IR. The debug info refers back to the LLVM-IR so that, for example,
in perf/vtune/gdb we have a map from assembly instructions back to
LLVM-IR. There are two options, insert_preopt_debug_info and
insert_postopt_debug_info. preopt inserts debug info before the module
is optimized and postopt inserts it after the module is optimized. If
both are set, postopt will override preopt. Both require
print_generated_code to be set, as the debug info must refer to a source
file. All options are off by default, as they will slow down the
compilation. (print_generated_code already defaulted to off).

This code derives from a PR to LLVM that introduced a pass to do this
(https://reviews.llvm.org/D40778), and a utility that is based on this
unmerged PR (https://github.com/vaivaswatha/debugir). The original
license is maintained for that code.