[build] Upgrade to jupiter-interface:0.9.0
net.aichler:jupiter-interface:0.7.0 can no longer be downloaded.
This breaks the planner build. This commit upgrades to 0.9.0 which is
available. This version seems to require adding jupiter-interface to the
test dependencies in addition to the sbt plugin.

A bunch of planner tests are still failing and the planner needs some
maintenance, but the same set of tests seem to be failing as before
this commit when we used 0.7.0 and 0.7.0 was still available for
download.