[olap] GRouter - add random data local policy
Add the DISTINCT_RANDOM_SPLIT_DATA_LOCAL policy.
This policy uses distinct sets of queues for each consumer. For a
random policy this doesn't matter very much, but if we want to adapt the
policy at runtime we will need consumer specific queues. This policy
is an example of how to have consumer specific queues.