[olap] Make SpecificCpuNumaNodeAffinitizer work with GeneralizedRouter
The largest change is that RandomSplitDataLocal uses Affinitizer
instead of AffinityPolicy. This is because we only need to find the
local CU/NUMA node, not map onto an array of [0, fanout) indexes like
we do in the standard Router. The previous approach of using
an AffinityPolicy with fanout = Affinitizer.size breaks down with
SpecificCpuNUMA as it can map onto a non-dense set of indexes.
 e.g SpecificCpuNumaNodeAffinitizer({2,4}) in a system with 8 numa
nodes.