[olap] improve GRouter open time with smaller AsyncQueueMPMC
GRouter has taking 100s of MS to open with multiple consumers and
DISTINCT_RANDOM_SPLIT_DATA_LOCAL since it had to create a AsyncQueueMPMC
per NUMA/GPU and per consumer. AsyncQueueMPMC involved a massive
allocation for its data member, and its not clear why it had to be so
large. it used N = 8k of a 64k sized vector. By reducing both N and the
vector size to 2K we get much faster creation time. I believe this
effectively reduces the maximum queue size to 2K, but we rarely use
slack above 128