[olap] Use shared ptrs for plugins in the catalog
With raw ptrs, there was no clear owner of a plugin and there were
leaks when the catalogue was cleared. There still is no real clear
owner as there are shared ptrs everywhere, but at least Plugin
destructors will be called.

For NVMe experiments this was problematic as if multiple experiments
were run in the same executable, the later experiments would be slower
than if they ran individually. Likely due to FDs being kept in the
NVMePlugin and closed with its destructor