[olap] add Operator ownership model
Operators were, in general, constructed with new and never deleted.
This is a problem for long-running experiments that run many prepared
queries. This commit makes parents own, or share ownership, of their
children through shared pointers. When the prepared query goes out of
scope the operator tree will now be freed, since the root operator is
held in the prepared query after the query is constructed through the
relbuilder. Children have non-owning pointers to their parents to avoid
ownership cycles.