[codegen] add type checking for binary expressions
There was already some limited type checking in ExpressionGeneratorVisitor,
but this occurs at JIT code generation time, so the error can be hard to
track back to the original c++ creating the expression_t objects, e.g.
in a query plan. If we use the wrong types, we often end up with LLVM
compiler errors instead, e.g. LLVM will complain about creating binary
instructions that compare different types.This can be quite confusing
and hard to debug the root cause for people not familiar with the
codegen infrastructure.

This commit adds some type checking in the creation of expression_t objects.
First, for all binary expressions, if the lhs is _not_ primitive, we
check LHS type == RHS type. We check if the lhs is not primitive because
sometimes it makes sense for binary ops of different types, e.g. IndexedSeq.
Second, we also add type checking in the comparison expressions, to have
slightly better error messages. Without adding a magic enum dependency,
we just have the opID when we check in the TBinaryExpression constructor