[olap] update metadata format for compression type
Before the metadata just contained COMPRESSED or UNCOMPRESSED in the
data_format field. This commit deprecates the data_format field and
replaces it with compression_type which corresponds to values in the
NvmePlugin::CompressionFormat_t enum.

This commit also exposes more info about the compression format in
NvmePlugin so that MemMoveDevice can set up its decompressors
appropriately. For now we require all attributes loaded by MemMoveDevice
to use the same format.